Day 5: Add variants that use HashSet and nom.
Using a HashSet instead of HashMap (and not having to compare against the value found in the HashMap) is notably faster.

Once again, I'm surprised that nom is every so slightly faster than manually splitting.  I like that the Rules parser returned the fully formed HashSet, without having to create a Vec or insert items manually one at a time.