Day 9: fix bugs in part2_heaps (works and FAST!)
The first bug was due to using Reverse() on the Position in the heaps.  It was finding the largest (not smallest) starting position from any of the heaps.  I had to use ".0" to access the inner value to get a comparison that was understandable.  (I guess I could have used another heap?).

See, this is why there ought to be a min-heap version.  And there ought to be a way to pass a compare function or key function to get the sorting you need, without having to wrap the items being stored.

The second bug is that if we found a free chunk of the appropriate size, we also need to check that it comes before the file.  That is, don't move a file to a larger starting location.