Day 21: Avoid the missing space on the directional pad
The logic for avoiding that missing space was backwards.  Added some sanity checks to detect when the robot is over that missing space (needs to be removed or turned into an assert!()).

Funnily enough, part 1 was still getting the correct overall answer, despite the specific sequence of movements being incorrect.

Alas, I can't tell if it now gets the correct answer for part 2, since it crashes before finishing.  I suspect the OS is killing it due to too much memory usage (perhaps for too long a period?).