Day 21: Trying a caching version of part 2.
It now tries to cache the number of key presses needed for particular sub-strings, and a given number of robots typing on directional keypads.

Added some more test cases.

My answers are now too high.  Unfortunately, I don't know if I'm generating the wrong keypresses, or if it's something wrong in the new implementation.  This is when I really want to be able to apply my brute force solution.