Day 21, part 2: slightly faster (5%)
We don't need to keep a HashMap of  sequence of changes to price for each buyer.  We just need a set to detect when we've seen a sequence of changes for the first time, at which point we just add the price to the HashMap that collects prices from all buyers.